app/requirements.txt
redis==5.0.8
pysyncobj==0.3.12
fastapi==0.115.0
uvicorn[standard]==0.30.6
numpy==1.26.4
Pillow>=10.0
